fix: mocking of sys.stdin and sys.stdout
